modifications

+++ src/main/java/com/gr4/api_gestion_budgets/repository/DepenseRepository.java
@@ -3,5 +3,5 @@ package com.gr4.api_gestion_budgets.repository;
 import com.gr4.api_gestion_budgets.models.Depense;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DepenseRepository extends JpaRepository<Depense, Long> {
+public interface DepenseRepository extends JpaRepository<Depense, Integer> {
 }
